Revert "Moved all tasks into one folder (#21)"
This reverts commit 274b9cc88ecf56587f637c230e116e9c0acf614f.
